Ticket #148 - Add prototype java jetty http server to JNLTest.java in jjnl as proof of concept
Add prototype java jetty http server to JNLTest.java in jjnl as proof of
concept.

The sample jnl_publisher.c posts a JAL-message of initialize in the http
header, which the java jetty subscriber reads and returns initialize-ack
in the response header which is displayed in the console of the
jnl_publisher c client.

The new subscriber java executable to build/run is JNLSubscriber.java
under jnl_test.

References #148