Ticket #540 - Fixed issue where subscriberImpl prepareSubscribe method wasn't checking for mode of archive for journal resume
Fixed issue where subscriberImpl prepareSubscribe method wasn't checking
for mode of archive for journal resume.

This resulted in any journal record being treated as a journal resume on
intialize and keeping the output directory when it should be deleted on
all journal records insteead of just keep the output directory on
archive mode only.

References #540