Ticket #580 - Fixed journal resume when publisher sends a record not matching the jalid of the record to be resumed
Changed so that if a journal resume is set on initialize and the
publisher sends a record with a different JAL id than the existing
partial record on the subscriber, then delete this partial record and
treat this upload as a new record instead of a resume.

Added a unit test to verify this change.

References #580