Ticket #612 - Subscriber rejects resends
Changed so if an invalid digest response message is received, then the
subscriber will force the digest status of invalid and force the
deletion of the temporary record so that it can be resent.

Added unit test to verify this.  Also trimmed back the number of record
posts in unit test to speed up build.

References #612